refactor(search): move search scenarios to `k6-utils` module
- Move the `searchScenarios` array from `search.ts` to `k6-utils.ts`
- Update the `Scenario` type to use `Omit` and `&` for better type safety
- Use `satisfies` operator to ensure `searchScenarios` matches `Scenario[]` type
- Update import statements in `search.ts` to import `searchScenarios` from `k6-utils.ts`
- Remove unused `Options` interface from `search.ts`
